Crate sem

Modules

ffi

Macros

ftok!
The ftok macro returns the System-V'IPC key from pathname.

semctl!
The semctl macro configures the semaphore index.

semctl_clear!
The semctl_clear macro removes the goup of semaphores.

semctl_init!
The semctl_init macro sets/inits the semaphore at 0.

semget!
The semget macro obtains a semaphore id.

semget_id!
The semget_id macro obtains a semaphore id and determinates the required flag.

semop!
The semop macro sets the semaphore index according to a SemBuffer.

semop_lock!
The semop_lock macro locks the semaphore.

semop_unlock!
The semop_lock macro unlocks the semaphore.